Better separation of docker-compose to start a fresh dev instances without all the functional tests with use
