+++ .ci/test-cluster.yml
@@ -4,7 +4,7 @@ x-kuzzle-config: &kuzzle-config
   image: kuzzleio/kuzzle-runner:${NODE_VERSION:-16}
   command: > 
     bash -c "
-      node ./docker/scripts/start-kuzzle-dev.js --enable-plugins kuzzle-plugin-cluster,functional-test-plugin
+      node ./docker/scripts/start-kuzzle-test.js --enable-plugins kuzzle-plugin-cluster,functional-test-plugin
     "
   volumes:
     - "..:/var/app"
